pyproject.toml
[project]
name = 'hooky'
description = 'Receive and respond to GitHub webhooks'
version = '1'
dependencies = [
    'asyncer',
    'cryptography',
    'fastapi',
    'httptools',
    'PyGithub',
    'PyJWT',
    'redis',
    'requests',
    'uvicorn',
    'uvloop',
    'rtoml',
    # to do remove once httpx updates `pip install -r requirements/testing.txt -r requirements/pyproject.txt`
    # no longer breaks
    'h11<=0.12.0'
]

[tool.flake8]
max_line_length = 120
max_complexity = 14
inline_quotes = 'single'
multiline_quotes = 'double'
ignore = ['E203', 'W503']

[tool.pytest.ini_options]
testpaths = 'tests'
log_format = '%(name)s %(levelname)s: %(message)s'
filterwarnings = [
    'error',
    'ignore:There is no current event loop:DeprecationWarning',
]
timeout = 30
xfail_strict = true

[tool.coverage.run]
source = ['src']
branch = true

[tool.coverage.report]
precision = 2
exclude_lines = [
    'pragma: no cover',
    'raise NotImplementedError',
    'raise NotImplemented',
    'if TYPE_CHECKING:',
    '@overload',
]

[tool.black]
color = true
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310']
skip-string-normalization = true
skip-magic-trailing-comma = true

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
color_output = true
